rtas_errd: Add support for hotplugging qemu pci devices via RTAS event
QEMU uses the newly defined hotplug event in librtasevent to initiate hotplug
of virtio pci devices in a pseries guest. This patch catches hotplug events and
invokes the drmgr commad with the necessary parameters to fufill the hotplug
request. Current support is limited to pci devices identified by a drc-index.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>
Acked-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>