Update extract_opal_dump utility for new sysfs interface
This patch reworks extract_opal_dump to use the revised sysfs interface
to retrieve and acknowledge platform dumps. It does not yet deal
with the new way to get error logs, that is a future patch to opal_errd.

We also add the ability to test the utility by pointing to a sysfs
like directory tree that isn't sysfs, we also add some test data
in sysfs-test/ that can be used to test the functionality of
extract_opal_dump.

The test-extract_opal_dump shell script will test the functionality of
the utility.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Acked-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>