opal_errd: update to use current sysfs interface
Update opal_errd to use current sysfs interface, using inotify to wait
for new events.

This is still bug-for-bug kind of compatible with the old sysfs interface
and does not fix various issues with writing to a single /var/log/platform
file with O_APPEND.

the test-opal_errd script can do a small test of extracting out
error logs from sysfs (well, a simulated sysfs). The test log there
is just all zeros as it turns out that works well enough for testing.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Acked-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>