opal_errd: fix regression in waiting for new log events.
This is a not-too-elegant fix for a regression where opal_errd would
not pick up on new elog events.

Basically, there are at least 3 ways of getting a notification from linux
that a "file" has changed/comes to exist:
1) poll() the fd of the sysfs file
2) inotify
3) udev

Only one of these is correct for any of the following scenarios:
- file/directory on normal filesystem
- sysfs file
- directory in sysfs

So instead of pulling in libudev as a dependency and then waiting on possibly
two file descriptors so that we work in both a test suite and on sysfs,
just add a select() call with a timeout.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>