opal-elog-dump: fix parsing of section headers
There were a couple of bugs in parsing section headers:
1) id being treated as uint16_t when in fact it is char[2] ascii
   according to docs
2) not endian safe
3) no error handling for invalid/corrupt headers
   (there's several hints we can look at).
4) header structure was not marked as packed
5) not all section header types handled

Note that this patch does not fix the endian issues in printing
the section contents, just parsing the header itself.

We use a new function to parse the header and check for corruption as
each section has one of those so we should reuse code.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>