opal-elog-parse: parse additional sections
Added datastructures and parsing code for SRC subsection 'FRU' and
all of the FRU subsections.

Additionally added checks to ensure that the required headers are
present within the log file, currently the parsing code only requires
a private header but user headers and extended headers should also be
present.

Basic FRU printing section was added but only barely enough to determine
that the parsing appears to work. This should be extended.

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>