test-opal-elog-parse: add tests for fru callout sections
No tests actually examine the output of code added in
f911af0409587e77d98aad6f9d7882ccea929237.

The current test is a truncated raw file from a real fsp dump.
The output was designed to be similar to the output from the fsp
but the test does not compare to real fsp.

[Fixed test-opal_errd.result - Vasant]

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>