opal-elog-parse: handle user header hidden action flag correctly
Original decisions were made to ignore the hidden action flag and simply
not report its presence or absence.

It has been decided that this flag should be renamed to a system health flag
and if it is present just report that the system is healthy and no service
action is required.

All mentions of HMC have been removed s/HMC/Hypervisor/

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>