extract_opal_dump: Pass sysfs path and maximum dumps
We are calling extract_opal_dump tool without any argument.
If we start opal_errd daemon with non default sysfs path (-s option)
then pass that info to extract_opal_dump tool as well so that it looks
for proper path and also pass maximum number of dumps of a specific
type to be saved.

[Reworked -m option passing - Vasant]

Signed-off-by: Aruna Balakrishnaiah <aruna@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>