opal-elog-parse: fix buffer overflow in parse_ch_scn
Current code mallocs the exact size of a string and uses strncpy
to copy the memory from the elog file. The value of n passed to
strncpy was the maximal posible value for the string and strncpy
was padding potentially large regions of memory with nulls.

Passing the correct n to strncpy along with some better error checking
beforehand solves this problem.

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>
Reviewed-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>