opal-elog-parse: fix potential bugs
When a parsing struct is malloced based on the headers length often the
length of the header is not examined which could result in mallocing an area
too small for the full struct. Often enough as well on variable length
sections the length of the variable section is not checked against what
the header length is.

Also found a places were the null from malloc isn't checked for.

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>
Reviewed-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>