opal-elog-parse: stop printing elog sections as soon as they are parsed
Now the code builds a opal_event_log during the parsing of an entire log and
subsequently prints an entire log once it has parsed the entire thing. The
positive of this is that it could now return this log stucture and do
various things with it later before printing it.

Signed-off-by: Cyril Bur <cyril@ozlabs.au.ibm.com>