diags: Keep man page source in git tree, not compressed file
Presently we are keeping compressed man page in source code which
is not good.This patch fixes this issue by keeping man page source
in git tree.

Also rename doc dir as man so that we follow consistent naming across
directories.

Note that there is no change in man page content.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>