lpd: Enable little endian platforms
This patch enables the LPD tool to work on little
endian platforms. After receiving the indices data from
rtas in big endian format, the location code structure
stores it in host format except the length and location
code string which are being used as input to rtas calls
and must be in big endian format.

Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>