convert_dt_node_props: LE support and cleanup
convert_dt_node_props command is used by the rtas_errd to
convert drc_index to name and vice-versa. The command reads
from the /proc/device-tree files which contain BE data, hence
need special care to convert the data to host format before
we process it.

NOTE: The thumb rule we follow is :
* Keep all the data used in the tools in host format.
* Convert anything we push-to/pull-from PAPAR sources.

Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>