rtas_errd: Add support for cpu and memory hotplugging via RTAS event
This patch adds support for handling RTAS hotplug events for cpus and memory
blocks in addition to such events for pci devices. It includes additional
support for events indicating an addition or a removal of multiple cpus or
logical memory blocks. Finally, for debugging purposes, it includes the option
to handle the event without calling the drmgr command.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Acked-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>