opal-dump-parse: Update man page for printk buffer
SYSDUMP contains raw kernel log buffer data. We need to know the
kernel log buffer internel structure (defined in kernel) to parse
this section. Adding those structure here will introduce dependency
on kernel and any change in that structure going to break this tool.

Instead for now we are keeping this section as is and use commands
like strings to parse this section..which gives readable output..
which is sufficient for developers.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Aruna Balakrishnaiah <aruna@linux.vnet.ibm.com>