opal-elog-parse: add a line of '-'s at the end of each log
This is a purely cosmetic patch which ensures that each log ends in a
|------------------------------------------------------------------------------|

Each section now prints its self as:

| Section |
|---------|
|Content  |
|---------|

With the the exception of the Private Header which also adds the bar on the
top. This is done so that if a section is printed on its own, the content is
framed, this avoids ambiguity as to whether or not a section has finished
printing.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>