rtas_errd: LE update for parsing drconf memory
When add phandles for memory in the ibm,dynamic-memory property, the first
32 bits of the property are the number of LMBs in the property. This
needs to be read as a BE value, failure to do this results in allocating
way more phandle structs that necessary and reading past the end of the
end of the buffer holding the property.

Add proper conversion for LE systems.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>