rtas_errd: Free phandle list after use
During PRRN event handling we build a list of phandles for all the LMBs
on the system in order to correlate between drc indexes and their
phandle. This list is never free'ed. Failure to do so results in a list
that grows every time a PRRN event is handled.

After handling a PRRN event free the list of phandles.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>