PowerKVM is a distro, not a platform
PowerKVM is a distro, not a platform. Bare metal systems are not
PowerKVM (which is IBM distro), but PowerNV.
This patch fixes naming convention for baremetal system.

Before: # rtas_errd
rtas_errd: is not supported on the PowerKVM platform
After: # rtas_errd
rtas_errd: is not supported on the PowerNV platform

Signed-off-by: Vaishnavi Bhat <vaishnavi@linux.vnet.ibm.com>
[Updated subject, description - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>