rtas_errd: Workaround for concurrent PRRN event collisions
Proposed workaround for issues with current PRRN event handling mechanism.

Currently, the mechanism works by kicking off a script that calls a drmgr
remove followed by a drmgr add for each effected device. With this method,
problems can occur when multiple prrn events arrive in a short time span or
when other dlpar operations are happening while the script is running. What
can happen is that these other dlpar operations can jump in inbetween when
the script calls the remove and the add and take the drmgr lock. Then when
we try to add the memory back from the script, we are unable to take the
drmgr lock and the add fails.

This patch will follow up a patch to drmgr which would add a new -P to
drmgr which would take the functionality of the existing prrn_hotplug
script and perform all the necessary removes/adds from within the drmgr
lock.

This patch to rtas_errd would simply construct a new file for each PRRN
event received containing all the name and drc index of each effected
device and pass that to the new drmgr option. When drmgr has completed
the operation or if it fails for any reason, the file will be deleted.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Acked-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
[Removed prrn_hotplug [un]install rules from Makefile - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>