diag/slider.c: Fix build error on Ubuntu
slider.c: In Ubuntu builds (14.04, 15.10, 16.04), there is an unfortunate interaction between
the build check '-Werror=format-security' and the format of the 'snprintf()' function.  In at
least one case, the optional arguments must be present to prevent generation of the error:

slider.c:149:3: error: format not a string literal and no format arguments [-Werror=format-security]

Addition of a dummy argument for the 'format' resolves the issue.

Signed-off-by: Michael W. Bringmann <mwb@linux.vnet.ibm.com>
[Added format specifier instead of NULL - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>