lpd: Add support for Marvell HDD LEDs on S822LC for HPC
This patch introduces support for the identification/fault LEDs of the
Marvell SATA HDD controller present in the IBM Power System S822LC for
HPC (a.k.a. Garrison/Minsky).

  indicator.h:        add the indicator type 'TYPE_MARVELL'.
  indicator_marvell.{h,c}:
                      implement the 'indicator' interface (get indices,
                      get/set indicator), and required hardware logic.
  indicator_opal.c:   call such interface on 'TYPE_MARVELL' devices.
  Makefile.c:         build and reference the indicator_marvell.o file.

The code for the hardware logic is based on a sample code written by
Douglas Miller <dougmill@us.ibm.com>, acknowledged in indicator_marvell.c.

Signed-off-by: Mauricio Faria de Oliveira <mauricfo@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>