lpd: marvell: handle non-unique/duplicate location codes
On some systems, it's been observed that different disks of the Marvell
SATA HDD controller might be listed with the very same location code.

This could cause an user to specify a device by location code, and have
the wrong device identified; if the service operation is a hotplug/pull
the disk, that error would cause serious problems.

Even though this might be a problem that requires a fix elsewhere, it's
still possible to support such scenario, with the usysident option that
specify the element by device name (e.g., '-d sda') instead of location
code (e.g., '-l -B0-T0-L0').

This patch adds a simple check and workaround for that -- restricted to
the Marvell devices (TYPE_MARVELL); once a device match is found:

- if the device is specified by device name,
  then proceed normally.

- if the device is specified by location code,
  then check for other devices with this location code.

  - if not found,
    then proceed normally.

  - else,
    stop, and ask the user to specify the device name with '-d'.

Signed-off-by: Mauricio Faria de Oliveira <mauricfo@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>