usysident: Add option to scan only the system indicators
usysident uses the lsvpd to get the location code and vpd info
for the enclosure devices. If there is a vpdupdate pending,
the lsvpd internally calls the vpdupdate and that can lead
to usysident command taking a lot of time.

Add an option to just print/update platform indicators and
skip all other indicators like enclosure indicators, disk
indicator, etc,.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>