Fix the spec file to work on SUSE
Many issues are fixed for building on SUSE.

The rpm names that the package depends on for build are different
on SUSE. So depending on the build host distro mark BuildRequires
correctly.

Also, the debug packages are not built by default on SUSE. Needs to
be explicitly state build so with a global attribute.

The SUSE marks the default for _libexecdir_ to /usr/lib. On Redhat
based distros, this variable defaults to /usr/libexec. Fix the
inconsistency by overwriting the _libexecdir_ to /usr/libexec on
SUSE.

While at this, also fix the whitespacing which are interchangeably
used at places with "spaces" and "tabs". Make all of them "tabs"
for consistency.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>