opal_errd: convert the old log type to new
The commit b797f66e6 changed the pruning logic to separate
counters for informational and serviceable log files. While
doing that it changed the filename to represent the log type.

The old existing log file names don't conform to the new
convention and the pruning logic would only prune the files
with the new naming convention format. So, the opal-log-parse
would always show the old log files which are never pruned
and the new ones are pruned in-between which makes the logic
appear broken.

The fix here is to rename the old files to new formatting
convention at daemon start so that, the pruning logic
would pick the files for pruning when required.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
[Reduced log severity, fixed possible descriptor close issue - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>