opal-elog-parse: Skip change of directory when file is given
c77991f8 started to handle the return values for all the system calls.
The chdir earlier called upon when the file path itself given used to pass
as the failure from that was not handled. With the mentioned commit, the
read_elog began to return failure when chdir failed, resulting in command
failure when -f is given.

This patch skips the chdir when the path is given by the user.

Fixes: c77991f8 (Fix compiler warning for return value being ignored)
Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>