diags/diag_nvme: Collect IBM VPD data from PCIe
Retrive VPD data from PCIe, old devices tend to use this mechanism, the
Log Page 0xF1 might not be supported, so have this option implemented as a
potential fallback.

The get_ibm_vpd_pcie funtion follows the PCI Express Base specification in
terms of how data should be parsed. The theorical maximum length of a
field is stored in 2 bytes (max 255).

On the other hand the set_vpd_pcie_field function uses a VPD structure
that matches the maximum lengths for PCIe VPD fields that are specified in
the NVMe IBM Purchase specification, that has maximum expected lengths
lower than the theoretical possible values in the PCIe spec.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>