diags/diag_nvme: Run diagnostics on specified NVMe device
This commit adds the initial functionality of diagnosing if SMART data can
be properly retrieved from an specified NVMe device, in case the data
can't be retrieved assume device is in a bad state and report an event in
the servicelog database.

Error events or higher are marked as serviceable events and marked as call
home candidates. Otherwise events will only be logged in the servicelog
database without serviceable or call home flags.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>