diags/diag_nvme: Avoid duplicate serviceable events
Query the servicelog database for any existent open serviceable events for
the NVMe device.

When an event already exists that require some action for a device in a
particular physical location code, do not generate a new event that would
trigger a call home event. When the existent serviceable event has been
repaired and closed, device is expected to be in an healthy state and new
error events can be logged again.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>