diags/diag_nvme: Report SMART data critical warning bit events
In case the SMART data has critical warning bits enabled report an event
in the servicelog, the only recoverable bit is the temperature one, if it
is the only one detected report as a warning event in the servicelog. If
any other bit is set to 1, bump the severity of the event to an error and
report all the bits set to 1 as part of the description.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>