diags/diag_nvme: Read SMART data from a file
Add an option to read the SMART data from a file instead of the device,
this can be used as a test mechanism, given injecting errors into a
device SMART log is not possible at this moment.

In case this option is used and the file pointed to can't be opened,
treat it as the failure to read SMART data from the device. Otherwise
process the SMART data contained in the file and use it for the
diagnostics of the device. In case failures are detected report the
events in the servicelog database.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>