diags/diag_nvme: Add extra data to logged event through raw_data
The raw_data field from the servicelog event allow to add extra
binary data to the logged event in the database. With it we can add
SMART and VPD data that was collected from the device which we are
generating the event.

The data itself is formatted as follows so it provides an extensible way
to add more data in the future if needed:

The raw_data format is made of sections, each section contains a
one-byte tag which identifies what is the data contained in this
section, an additional one-byte reserved and two-byte length which
represents the length value (in bytes) of all the data fields contained
in this section, followed by the data items, so a section has the
following format:
+------------+----------------------------------------------+
|   Offset   |                Field meaning                 |
+------------+----------------------------------------------+
|   Byte 0   |               Identifier Tag                 |
|            |                                              |
|   Byte 1   |                  Reserved                    |
|            |                                              |
|   Byte 2   | Length (in bytes) of data - bits[7:0] (LSB)  |
|            |                                              |
|   Byte 3   | Length (in bytes) of data - bits[15:8] (MSB) |
|            |                                              |
| Byte 4 - N |                 Data items                   |
+------------+----------------------------------------------+

The current tags in use are the following:
+---------+------------+
| Tag 01h | SMART data |
+---------+------------+
| Tag 02h |  VPD data  |
+---------+------------+
| Tag FFh |  End Tag   |
+---------+------------+
For the end tag (FFh) no additional data items are present along with
the header, so it is a section with other header bytes set to 0.

Each data item consists of a four-byte header followed by some amount of
data. The format of this header is as follows: a keyword is a
two-character (ASCII) mnemonic that uniquely identifies the information
in the field, one additional reserved byte and the last byte of the
header is binary and represents the length value (in bytes) of the data
that follows. The keyword data is provided as ASCII characters. The
following is an example:
         |      +0       |      +1       |      +2       |      +3       |
         |               |               |               |               |
         |7|6|5|4|3|2|1|0|7|6|5|4|3|2|1|0|7|6|5|4|3|2|1|0|7|6|5|4|3|2|1|0|
         +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
         |            Keyword            |   Reserved    |   Length (N)  |
Byte 0 ->|                               |               |               |
         +-------------------------------+---------------+---------------+
         |               Data (byte 4 -> N+3, example here N=4)          |
Byte 4 ->|                                                               |
         +---------------------------------------------------------------+

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>