common/platform: Add function to identify processor version
Add a function to identify what processor version is in the system. This
is done by looking into the Processor Version Register (PVR) first two
bytes.

The PVR is a more reliable way to detect this given processor in compat
mode would display a different processor version, but still label the
proper PVR.

For example a P10 processor running in p8compat mode:
processor       : 0
cpu             : POWER8 (architected), altivec supported
clock           : 3600.000000MHz
revision        : 2.0 (pvr 0080 0200)

As opposed to what it would be displayed if it was not in compat mode:
processor       : 0
cpu             : POWER10 (architected), altivec supported
clock           : 3600.000000MHz
revision        : 2.0 (pvr 0080 0200)

So use this information to properly label which processor generation is
in use.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>