diags/diag_nvme: Enable diag_nvme build and limit its scope
Add makefile changes required to build the diag_nvme code and a man page
describing the command usage. Also add a script to run the diag_nvme
tool and install it to run as a cron daily job.

The scope in which this tool should be executed is limited to P10 Power
VM LPARs.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>