diags/diag_nvme: Trim trailing white space in sl_callout fields
The string fields in sl_callout struct can't have white spaces or it
will break the current way the servevent_parse.pl script works in which
the input is split using a space separator. So any extra white space
will cause the fields to move around.

If IBM VPD data is extracted from NVMe log page 0xF1 it is right padded
with white spaces, so the strings needs to be trimmed.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>