scripts/ppc64_diag_mkrsrc: Fix partition ID for LE OS releases
When running the 'od' command to retrieve the partition ID from the
device tree on a litte-endian OS the result isn't properly displayed.
For example on a partition with ID 3 the following is seen:

$ od -An -td /proc/device-tree/ibm,partition-no | sed "s/\\s//g" 2>/dev/null
50331648

The result is 50331648 (0x03000000) when the expected is actually 3
(0x00000003). This issue is not seen in big-endian OS releases where the
value retrieved is properly displayed. So add --endian=big flag to
address this issue for LE releases and keep the same behavior seen in BE
releases:

$ od -An -td --endian=big /proc/device-tree/ibm,partition-no | sed "s/\\s//g" 2>/dev/null
3

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>