rtas_errd: Fix core dump with large filename passed to -f option
When a filename passed to rtas_errd's -f option is excessively long,
it causes a stack buffer overflow in the _log_msg() function, leading
to stack smashing.

The _log_msg() function, utilized by log_msg(), operates with a buffer
limited to 4096 characters. Additional space is consumed by _dbg() and
re-formatting in _log_msg(), increasing the risk of buffer overflow.

To address this, in case of an error, if the filename exceeds the
length of 2048 characters, truncate it and log a message indicating the
truncation. Use 2048 characters as limit for filename length to allow
space for further modifications.

Additionally, add a check when processing the -f option to ensure that
the length of the specified filename does not exceed 4096 characters.

Reported-by: R Nageswara Sastry <rnsastry@linux.ibm.com>
Closes: https://github.com/power-ras/ppc64-diag/issues/31
Signed-off-by: Sathvika Vasireddy <sv@linux.ibm.com>
Tested-by: R Nageswara Sastry<rnsastry@linux.ibm.com>
Signed-off-by: Mahesh Salgaonkar <mahesh@linux.ibm.com>